requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-countries==7.6.1
django-environ==0.12.0
django-multiselectfield==0.1.13
django-phonenumber-field==8.1.0
django-unfold==0.54.0
gunicorn==23.0.0
idna==3.10
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pytz==2025.2
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
